Major cleanup to make libs, examples, and other folders pass all linters

+++ src/libs/test-libs/utillib/exportContext/testing.cpp
@@ -10,8 +10,9 @@
  * General Public License for more details. You should have received a copy of the GNU General
  * Public License along with this program. If not, see http://www.gnu.org/licenses/.
  *-------------------------------------------------------------------------------------------*/
-#include "utillib.h"
 #include <iomanip>
+#include <string>
+#include "utillib.h"
 
 #define sizeme(a) ((uint64_t)sizeof((a)))
 //--------------------------------------------------------------
@@ -29,8 +30,8 @@ int main(int argc, const char *argv[]) {
     unsigned int ui = (unsigned int)-1;
     int i = -1;
     int64_t il = -1;
-    uint64_t uil = (uint64_t)-1;
-    float f = (float)10.333;
+    uint64_t uil = static_cast<uint64_t>(-1);
+    float f = 10.333f;
     double d = 1001001010.01121112;
 
     CStringExportContext ctx;
